refactor: StudyId Long으로 타입 수정

+++ android/app/src/main/java/com/created/team201/presentation/studyThread/ThreadActivity.kt
@@ -42,7 +42,7 @@ class ThreadActivity : BindingActivity<ActivityThreadBinding>(R.layout.activity_
     companion object {
         private const val STUDY_ID = "STUDY_ID"
 
-        fun getIntent(context: Context, studyId: Int): Intent =
+        fun getIntent(context: Context, studyId: Long): Intent =
             Intent(context, ThreadActivity::class.java).apply {
                 putExtra(STUDY_ID, studyId)
             }
